requirements-test.txt
git+https://github.com/zparis99/ECoG-foundation-model.git@v0.1.0#egg=ecog_foundation_model
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
einops==0.8.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mne==1.9.0
mne-bids==0.16.0
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==5.29.4
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
setuptools==78.1.0
six==1.17.0
smart-open==7.1.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
pytest==8.3.4
pytest-mock==3.12.0